add validation

+++ RestApiApp/service/src/main/java/com/epam/esm/service/impl/GiftCertificateServiceImpl.java
@@ -2,13 +2,12 @@ package com.epam.esm.service.impl;
 
 import com.epam.esm.entity.Criteria;
 import com.epam.esm.entity.GiftCertificate;
-import com.epam.esm.exception.DataNotExistRepositoryException;
-import com.epam.esm.exception.RepositoryException;
+import com.epam.esm.exception.*;
 import com.epam.esm.repository.GiftCertificateRepository;
 import com.epam.esm.service.GiftCertificateService;
-import com.epam.esm.service.exception.NotFoundServiceException;
-import com.epam.esm.service.exception.ServiceException;
+import com.epam.esm.validator.ServiceValidator;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -17,13 +16,19 @@ import java.util.Optional;
 @Service
 public class GiftCertificateServiceImpl implements GiftCertificateService {
 
+
+    private static final String INCORRECT_CERTIFICATE_MSG = "Incorrect certificate data";
     private final GiftCertificateRepository giftCertificateRepository;
 
+    private final ServiceValidator<GiftCertificate, Integer> validator;
+
+    private static final String INVALID_ID_MSG = "Certificate id is invalid";
     private static final String NOT_EXIST_MSG = "Gift Certificate id with number %s doesn't exist";
 
     @Autowired
-    public GiftCertificateServiceImpl(GiftCertificateRepository giftCertificateRepository) {
+    public GiftCertificateServiceImpl(GiftCertificateRepository giftCertificateRepository, ServiceValidator<GiftCertificate, Integer> validator) {
         this.giftCertificateRepository = giftCertificateRepository;
+        this.validator = validator;
     }
 
     @Override
@@ -34,6 +39,10 @@ public class GiftCertificateServiceImpl implements GiftCertificateService {
 
     @Override
     public GiftCertificate getById(int id) throws ServiceException{
+        if (!validator.isIdValid(id)){
+            throw new IncorrectDataServiceException(INVALID_ID_MSG);
+        }
+
         GiftCertificate giftCertificate;
         try {
             giftCertificate = giftCertificateRepository.getById(id);
@@ -47,6 +56,10 @@ public class GiftCertificateServiceImpl implements GiftCertificateService {
 
     @Override
     public GiftCertificate add(GiftCertificate giftCertificate) throws ServiceException {
+        if (!validator.validate(giftCertificate)){
+            throw new IncorrectDataServiceException(INCORRECT_CERTIFICATE_MSG);
+        }
+
         GiftCertificate resultCertificate;
         try {
            resultCertificate = giftCertificateRepository.add(giftCertificate);
@@ -58,6 +71,10 @@ public class GiftCertificateServiceImpl implements GiftCertificateService {
 
     @Override
     public void delete(int id) throws ServiceException {
+        if (!validator.isIdValid(id)){
+            throw new IncorrectDataServiceException(INVALID_ID_MSG);
+        }
+
         try {
             giftCertificateRepository.delete(id);
         } catch (DataNotExistRepositoryException e) {
@@ -69,6 +86,10 @@ public class GiftCertificateServiceImpl implements GiftCertificateService {
 
     @Override
     public GiftCertificate update(GiftCertificate current, GiftCertificate modified) throws ServiceException {
+        if (!(validator.validate(current) && validator.validate(modified))){
+            throw new IncorrectDataServiceException(INCORRECT_CERTIFICATE_MSG);
+        }
+
         GiftCertificate resultCertificate;
         try {
             resultCertificate = giftCertificateRepository.update(current, modified);
