+++ RestApiApp/service/src/main/java/com/epam/esm/service/impl/TagServiceImpl.java
@@ -1,14 +1,10 @@
 package com.epam.esm.service.impl;
 
 import com.epam.esm.entity.Tag;
-import com.epam.esm.exception.DataAlreadyExistRepositoryException;
-import com.epam.esm.exception.DataNotExistRepositoryException;
-import com.epam.esm.exception.RepositoryException;
+import com.epam.esm.exception.*;
 import com.epam.esm.repository.TagRepository;
 import com.epam.esm.service.TagService;
-import com.epam.esm.service.exception.AlreadyExistServiceException;
-import com.epam.esm.service.exception.NotFoundServiceException;
-import com.epam.esm.service.exception.ServiceException;
+import com.epam.esm.validator.ServiceValidator;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -17,14 +13,19 @@ import java.util.List;
 @Service
 public class TagServiceImpl implements TagService {
 
+    private static final String INVALID_ID_MSG = "Tag id is invalid";
+    private static final String INCORRECT_TAG_MSG = "Tag data is incorrect";
     private final TagRepository tagRepository;
+    private final ServiceValidator<Tag, Integer> validator;
 
     private static final String NOT_EXIST_MSG = "Tag id with number %s doesn't exist";
     private static final String ALREADY_EXIST_MSG = "Tag with name %s already exists";
 
+
     @Autowired
-    public TagServiceImpl(TagRepository tagRepository) {
+    public TagServiceImpl(TagRepository tagRepository, ServiceValidator<Tag, Integer> validator) {
         this.tagRepository = tagRepository;
+        this.validator = validator;
     }
 
     @Override
@@ -34,6 +35,9 @@ public class TagServiceImpl implements TagService {
 
     @Override
     public Tag getById(int id) throws ServiceException {
+        if (!validator.isIdValid(id)){
+            throw new IncorrectDataServiceException(INVALID_ID_MSG);
+        }
         Tag tag;
         try {
             tag = tagRepository.getById(id);
@@ -47,6 +51,9 @@ public class TagServiceImpl implements TagService {
 
     @Override
     public void add(Tag tag) throws ServiceException {
+        if (!validator.validate(tag)){
+            throw new IncorrectDataServiceException(INCORRECT_TAG_MSG);
+        }
         try {
             tagRepository.add(tag);
         } catch (DataAlreadyExistRepositoryException e) {
@@ -58,6 +65,9 @@ public class TagServiceImpl implements TagService {
 
     @Override
     public void delete(int id) throws ServiceException {
+        if (!validator.isIdValid(id)){
+            throw new IncorrectDataServiceException(INVALID_ID_MSG);
+        }
         try {
            tagRepository.delete(id);
         } catch (DataNotExistRepositoryException e) {
